Former-commit-id: ee2ec4fdc94e90b7fc91426deeae9820f227d9ab

+++ assets/shaders/simple_model_shader.vert
@@ -3,7 +3,7 @@
 layout(location = 0)in vec3 position;
 layout(location = 1)in vec3 normal;
 layout(location = 2)in vec2 uv;
-layout(location = 3)in ivec4 boneIndices;
+layout(location = 3)in vec4 boneIndices;
 layout(location = 4)in vec4 weights;
 
 uniform mat4 normalMatrix;
